feature: fail the analysis when we encounter a compile error
when we analyse a project it might happen that there's a compile error
and in that case we fail to find issues. If we complete the analysis we
report no issue to the SonarQube server and any existing issue is
therfore marked as solved.
Instead throw an exception and let the analysis fail